split derive code into modules
