+++ ff_derive/src/util.rs
@@ -1,0 +1,131 @@
+use num_bigint::BigUint;
+use num_integer::Integer;
+use num_traits::{One, ToPrimitive, Zero};
+
+/// Fetches the ident being wrapped by the type we're deriving.
+pub fn fetch_wrapped_ident(body: &syn::Data) -> Option<syn::Ident> {
+    match body {
+        &syn::Data::Struct(ref variant_data) => match variant_data.fields {
+            syn::Fields::Unnamed(ref fields) => {
+                if fields.unnamed.len() == 1 {
+                    match fields.unnamed[0].ty {
+                        syn::Type::Path(ref path) => {
+                            if path.path.segments.len() == 1 {
+                                return Some(path.path.segments[0].ident.clone());
+                            }
+                        }
+                        _ => {}
+                    }
+                }
+            }
+            _ => {}
+        },
+        _ => {}
+    };
+
+    None
+}
+
+/// Fetch an attribute string from the derived struct.
+pub fn fetch_attr(name: &str, attrs: &[syn::Attribute]) -> Option<String> {
+    for attr in attrs {
+        if let Ok(meta) = attr.parse_meta() {
+            match meta {
+                syn::Meta::NameValue(nv) => {
+                    if nv.path.get_ident().map(|i| i.to_string()) == Some(name.to_string()) {
+                        match nv.lit {
+                            syn::Lit::Str(ref s) => return Some(s.value()),
+                            _ => {
+                                panic!("attribute {} should be a string", name);
+                            }
+                        }
+                    }
+                }
+                _ => {
+                    panic!("attribute {} should be a string", name);
+                }
+            }
+        }
+    }
+
+    None
+}
+
+/// Convert BigUint into a vector of 64-bit limbs.
+pub fn biguint_to_real_u64_vec(mut v: BigUint, limbs: usize) -> Vec<u64> {
+    let m = BigUint::one() << 64;
+    let mut ret = vec![];
+
+    while v > BigUint::zero() {
+        ret.push((&v % &m).to_u64().unwrap());
+        v = v >> 64;
+    }
+
+    while ret.len() < limbs {
+        ret.push(0);
+    }
+
+    assert!(ret.len() == limbs);
+
+    ret
+}
+
+/// Convert BigUint into a tokenized vector of 64-bit limbs.
+pub fn biguint_to_u64_vec(v: BigUint, limbs: usize) -> proc_macro2::TokenStream {
+    let ret = biguint_to_real_u64_vec(v, limbs);
+    quote!([#(#ret,)*])
+}
+
+pub fn biguint_num_bits(mut v: BigUint) -> u32 {
+    let mut bits = 0;
+
+    while v != BigUint::zero() {
+        v = v >> 1;
+        bits += 1;
+    }
+
+    bits
+}
+
+/// BigUint modular exponentiation by square-and-multiply.
+pub fn exp(base: BigUint, exp: &BigUint, modulus: &BigUint) -> BigUint {
+    let mut ret = BigUint::one();
+
+    for i in exp
+        .to_bytes_be()
+        .into_iter()
+        .flat_map(|x| (0..8).rev().map(move |i| (x >> i).is_odd()))
+    {
+        ret = (&ret * &ret) % modulus;
+        if i {
+            ret = (ret * &base) % modulus;
+        }
+    }
+
+    ret
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    use std::str::FromStr;
+
+    #[test]
+    fn test_exp() {
+        assert_eq!(
+            exp(
+                BigUint::from_str("4398572349857239485729348572983472345").unwrap(),
+                &BigUint::from_str("5489673498567349856734895").unwrap(),
+                &BigUint::from_str(
+                    "52435875175126190479447740508185965837690552500527637822603658699938581184513"
+                )
+                .unwrap()
+            ),
+            BigUint::from_str(
+                "4371221214068404307866768905142520595925044802278091865033317963560480051536"
+            )
+            .unwrap()
+        );
+    }
+}
